[기능 수정] 로딩 컴포넌트 활성화, 게시글 코드 스플리팅
