+++ frontend/src/App.tsx
@@ -2,12 +2,9 @@ import { Suspense, useEffect, lazy } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import { useSelector } from './modules';
 import useLocalStroeageData from './common/hooks/useLocalStorageData';
-import CreateNotice from './components/Board/CreateNotice/CreateNotice';
 import GuestPage from './components/GuestPage/GuestPage';
 import Header from './components/Header/Header';
 import Home from './components/Home/Home';
-import ViewNotice from './components/Board/ViewNotice/ViewNotice';
-import EditNotice from './components/Board/EditNotice/EditNotice';
 import SuspenseLoading from './components/Loading/SuspenseLoading';
 import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
 
@@ -15,6 +12,7 @@ import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary';
 const UserPage = lazy(() => import('./components/User/UserPage'));
 const Crew = lazy(() => import('./components/Crew/Crew'));
 const CrewList = lazy(() => import('./components/Crew/CrewListPage'));
+const Board = lazy(() => import('./components/Board/Board'));
 
 function App() {
   const token = useSelector((state) => state.signIn.token);
@@ -29,7 +27,7 @@ function App() {
     <>
       <BrowserRouter>
         <ErrorBoundary>
-          <Suspense fallback={SuspenseLoading}>
+          <Suspense fallback={<SuspenseLoading />}>
             <Header />
             <Switch>
               <Route exact path="/" component={Home} />
@@ -37,9 +35,7 @@ function App() {
               <Route path="/user/" component={UserPage} />
               <Route path="/crewList" component={CrewList} />
               <Route path="/crew" component={Crew} />
-              <Route exact path="/boards/run/:runId" component={ViewNotice} />
-              <Route exact path="/boards/edit/run/:id" component={EditNotice} />
-              <Route exact path="/boards/create/run" component={CreateNotice} />
+              <Route path="/boards" component={Board} />
             </Switch>
             <div id="modal" />
           </Suspense>
